+++ challenge_template/solutions/UserName.py
@@ -4,4 +4,8 @@ def user_solution():
 def other_user_solution():
 	pass
 
-SOLUTIONS = [user_solution, other_user_solution]
+SOLVERS = [user_solution, other_user_solution]
+
+if __name__ == '__main__':
+	from app import run_solution_directly
+	run_solution_directly(__file__)