+++ challenges/01/solutions/RascalTwo.py
@@ -9,4 +9,8 @@ def smartest(nums: List[int], target: int) -> Optional[Tuple[int, int]]:
 
 one_liner = lambda nums, target: next(((x, nums[j]) for i, x in enumerate(nums) for j in range(i + 1, len(nums)) if i != j and x * nums[j] == target), None)
 
-SOLUTIONS = [smartest, one_liner]
+SOLVERS = [smartest, one_liner]
+
+if __name__ == '__main__':
+	from app import run_solution_directly
+	run_solution_directly(__file__)