+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # ChangeLog
 
+## 1.0.1
+
+### Added
+
+- `README.md` instructions:
+  - Running
+  - Adding Solutions
+- Direct-solution execution
+
+### Changed
+
+- `__main__.py` content moved to `app.py`
+- Renamed `SOLUTIONS` to `SOLVERS`
+
 ## 1.0.0
 
 ### Added